Several changes to new_comment_marker
The new comment marker has ben modified to highlight replies to your
posts in red instead of blue, and a separate count for reples to your
posts (if any) is displayed next to the filters button/link.

If there are replies to your post, the new count of replies to your
posts becomes clickable, which opens a 'dialog' (similar to the filters
'dialog'), with 'quick links' to the replies which if clicked, open
that post and scrolls to it.

The images used for the connecting lines between a post and it's replies
have been updated in two ways.  First, they now have a transparent
background, and the normal line color has been changed to medium gray
(#808080). This allows them to work better with scripts / styles that
change the background (the same image is now being used for normal and
white shack). This also allows them to work with scripts that add a
background image such as TroZ's ShackChattyNews. This has the side effect
of making the 'fade out' of the last few pixels to black / white not
feasable and still be a gif, so these were removed, which allowed the
image to be implemented with a 2 color palete, which slightly reduced
their size.

Finally, some improvements were made so the script better handles news
article pages.  The script now correctly detects the 'See All Comments'
link in the 'chatty_settings' area and no longer modifies it.

DETECTION OF YOUR POSTS
This is an important part of the script, as improperly doing this would
defeat the purpose of this modification.  There are two methods at the
beginning of th emodified script for accomplishing this, one of which
is currently setup to be called.  The first method checks local storage
first for a saved 'user id', and will assume it is correct if found.
Only if it is not found will it use an XPATH request to search for an
'a' tag with a 'this_user' class, and look at the parent node's classes
for the user's id.  The other method not currently being used takes the
opposite approach, always using an xpath request, and only used the
stored version if the XPATH come up empty. Both version store the id if
it is found when using the XPATH query.  The first version operates more
quickkly, but assumes one shacker per computer. The second (currently
unused) version is slower, but works better for shackers who sometimes
share a computer, correctly switching users to highligh after they post
at least once after logging in.